chore(lint): `golangic-lint`: fix various

+++ data/table/sheet/coordinates.go
@@ -10,13 +10,13 @@ const (
 	ExcelMaxColCount = 16384
 	ExcelMaxRowCount = 1048576
 	Alphabet         = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-	ZZ               = uint(702)
-	ZZIndex          = uint(701)
+	ZZ               = uint32(702)
+	ZZIndex          = uint32(701)
 )
 
 var alphabetSlice = strings.Split(Alphabet, "")
 
-func quotient26ToPrefix(quotient uint) string {
+func quotient26ToPrefix(quotient uint32) string {
 	// " ABCDEF"
 	// "0123456"
 	// Max Quotient =26
@@ -44,7 +44,7 @@ func quotient25ToPrefix(quotient uint32) string {
 }
 */
 
-func remainder26ToSuffix(remainder uint) string {
+func remainder26ToSuffix(remainder uint32) string {
 	if remainder == 0 {
 		return "Z"
 	}
@@ -55,7 +55,7 @@ func remainder26ToSuffix(remainder uint) string {
 	return letter
 }
 
-func remainder25ToSuffix(remainder uint) string {
+func remainder25ToSuffix(remainder uint32) string {
 	if remainder > 25 {
 		panic(fmt.Sprintf("remainder out of range: [%v]", remainder))
 	}
@@ -63,51 +63,51 @@ func remainder25ToSuffix(remainder uint) string {
 	return letter
 }
 
-func ColNumberToLetters(colNumber uint) string {
+func ColNumberToLetters(colNumber uint32) string {
 	if colNumber == 0 {
 		panic("row cannot be zero. Row is 1 indexed.")
 	} else if colNumber > ZZ {
 		panic("colNumber greater than 702 (ZZ) is not currently supported")
 	}
-	quotient := int64(0)
+	quotient := uint32(0)
 	if colNumber > 26 {
-		quotient = int64((colNumber - 1) / 26)
+		quotient = (colNumber - 1) / 26
 	}
-	remainder := int64((colNumber) % 26)
+	remainder := colNumber % 26
 	if colNumber < 26 {
-		remainder = int64(colNumber)
+		remainder = colNumber
 	}
-	prefix := quotient26ToPrefix(uint(quotient))
-	suffix := remainder26ToSuffix(uint(remainder))
+	prefix := quotient26ToPrefix(quotient)
+	suffix := remainder26ToSuffix(remainder)
 	return strings.TrimSpace(prefix) + strings.TrimSpace(suffix)
 }
 
-func ColIndexToLetters(colIndex uint) string {
+func ColIndexToLetters(colIndex uint32) string {
 	if colIndex > ZZIndex {
 		panic(
 			fmt.Sprintf("colIndex [%v] greater than index:701 (ZZ) is not currently supported",
 				colIndex))
 	}
-	quotient := int64(0)
+	quotient := uint32(0)
 	if colIndex > 25 {
-		quotient = int64((colIndex) / 26)
+		quotient = colIndex / 26
 	}
-	remainder := int64((colIndex) % 26)
-	prefix := quotient26ToPrefix(uint(quotient))
-	suffix := remainder25ToSuffix(uint(remainder))
+	remainder := colIndex % 26
+	prefix := quotient26ToPrefix(quotient)
+	suffix := remainder25ToSuffix(remainder)
 	return strings.TrimSpace(prefix) + strings.TrimSpace(suffix)
 }
 
 // CoordinatesToSheetLocation converts x, y integer coordinates
 // to a spreadsheet location such as "AA1" for col 27, row 1.
-func CoordinatesToSheetLocation(colIdx, rowIdx uint) string {
+func CoordinatesToSheetLocation(colIdx, rowIdx uint32) string {
 	colLet := ColIndexToLetters(colIdx)
 	return colLet + strconv.Itoa(int(rowIdx+1))
 }
 
 // CoordinateNumbersToSheetLocation converts x, y integer coordinates
 // to a spreadsheet location such as "AA1" for col 27, row 1.
-func CoordinateNumbersToSheetLocation(colNum, rowNum uint) string {
+func CoordinateNumbersToSheetLocation(colNum, rowNum uint32) string {
 	return CoordinatesToSheetLocation(colNum-1, rowNum-1)
 	//colLet := ColNumberToLetters(colNum)
 	//return colLet + strconv.Itoa(int(rowNum))
