+++ data/table/table_set.go
@@ -113,7 +113,7 @@ func (ts *TableSet) WriteXLSX(filename string) error {
 // ReadTableSetXLSXFile reads in an entire XLSX file as a `TableSet`. Warning: this can be resource
 // intensive if there's a lot of data. If you just want one sheet of many, it is better to
 // extract an individual sheet or sheets from an `excelize.File`, such as using `ParseTableXLSX()` or `ParseTableXLSXIndex()`.
-func ReadTableSetXLSXFile(filename string, headerRowCount uint, trimSpace bool) (*TableSet, error) {
+func ReadTableSetXLSXFile(filename string, headerRowCount uint32, trimSpace bool) (*TableSet, error) {
 	ts := NewTableSet("")
 	xm, err := excelizeutil.NewFile(filename)
 	if err != nil {
@@ -133,7 +133,7 @@ func ReadTableSetXLSXFile(filename string, headerRowCount uint, trimSpace bool) 
 	return ts, xm.Close()
 }
 
-func ReadTableXLSXFile(filename, sheetName string, headerRowCount uint, trimSpace bool) (*Table, error) {
+func ReadTableXLSXFile(filename, sheetName string, headerRowCount uint32, trimSpace bool) (*Table, error) {
 	if xf, err := excelizeutil.ReadFile(filename); err != nil {
 		return nil, err
 	} else {
@@ -149,7 +149,7 @@ func ReadTableXLSXIndexFile(filename string, sheetIdx, headerRowCount uint32, tr
 	}
 }
 
-func ParseTableXLSX(f *excelize.File, sheetName string, headerRowCount uint, trimSpace bool) (*Table, error) {
+func ParseTableXLSX(f *excelize.File, sheetName string, headerRowCount uint32, trimSpace bool) (*Table, error) {
 	if f == nil {
 		return nil, excelizeutil.ErrExcelizeFileCannotBeNil
 	}
@@ -169,7 +169,7 @@ func ParseTableXLSXIndex(f *excelize.File, sheetIdx, headerRowCount uint32, trim
 		return nil, excelizeutil.ErrExcelizeFileCannotBeNil
 	}
 	xm := excelizeutil.File{File: f}
-	if cols, rows, err := xm.TableDataIndex(uint(sheetIdx), uint(headerRowCount), trimSpace, false); err != nil {
+	if cols, rows, err := xm.TableDataIndex(sheetIdx, headerRowCount, trimSpace, false); err != nil {
 		return nil, err
 	} else {
 		tbl := NewTable(f.GetSheetName(int(sheetIdx)))
